auto-generate fill function with script data inside
